% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crainFeat.R
\name{crainFeat}
\alias{crainFeat}
\title{Shiny app to process example NP reports}
\usage{
crainFeat(
  md_dest = tempfile(),
  mods = c("gpt-4o-mini", "deepseek-r1:14b", "llama3.1:8b", "llama3.2-vision:latest")
)
}
\arguments{
\item{md_dest}{character(1), path to file to receive markdown from LLM}

\item{mods}{character() vector of model names useful with wizrd}
}
\description{
Shiny app to process example NP reports
}
