% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmdf}
\alias{nmdf}
\title{a simple data.frame with the ASDPTO vocabulary from 
bioontologies.org}
\format{
data.frame
}
\usage{
nmdf
}
\description{
a simple data.frame with the ASDPTO vocabulary from 
bioontologies.org
}
\examples{
data(nmdf)
head(nmdf)
}
\keyword{datasets}
