% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdf.R
\name{parse_nppdf}
\alias{parse_nppdf}
\title{parse text and extract tables from a pdf}
\usage{
parse_nppdf(pdfpath)
}
\description{
parse text and extract tables from a pdf
}
\note{
There can be tabulapdf errors thrown that do not register in R.
They do not seem to be trappable, so be sure you are comfortable with
the event noted.
}
\examples{
nelsonasd = system.file("pdfs", "nelsonAutDeid.pdf", package="llm4np")
nels = parse_nppdf(nelsonasd)
nels
}
